+++ artcoded/src/main/java/tech/artcoded/websitev2/notification/NotificationCleanupAction.java
@@ -4,6 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Component;
 import tech.artcoded.websitev2.action.*;
+import tech.artcoded.websitev2.utils.dto.Tuple.Tuple2;
 
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
@@ -23,20 +24,35 @@ public class NotificationCleanupAction implements Action {
     }
 
     @Override
+    public boolean shouldNotRun(List<ActionParameter> parameters) {
+        if (getFromParam(parameters) instanceof Tuple2<Long, Date>(Long _, Date searchDate)) {
+            return repository.countBySeenIsTrueAndReceivedDateBefore(searchDate) == 0;
+        } else {
+            return true;
+        }
+
+    }
+
+    Tuple2<Long, Date> getFromParam(List<ActionParameter> parameters) {
+        Long daysBefore = parameters.stream().filter(p -> PARAMETER_DAYS_BEFORE.equals(p.getKey()))
+                .filter(p -> StringUtils.isNotEmpty(p.getValue())).findFirst()
+                .flatMap(p -> p.getParameterType().castLong(p.getValue())).orElse(6L);
+        Date date = Date.from(ZonedDateTime.now().minusDays(daysBefore).toInstant());
+        return new Tuple2<>(daysBefore, date);
+    }
+
+    @Override
     public ActionResult run(List<ActionParameter> parameters) {
         var resultBuilder = this.actionResultBuilder(parameters);
 
         List<String> messages = new ArrayList<>();
         messages.add("before cleaning, count %s".formatted(repository.count()));
         try {
-            Long daysBefore = parameters.stream().filter(p -> PARAMETER_DAYS_BEFORE.equals(p.getKey()))
-                    .filter(p -> StringUtils.isNotEmpty(p.getValue())).findFirst()
-                    .flatMap(p -> p.getParameterType().castLong(p.getValue())).orElse(6L);
-            Date date = Date.from(ZonedDateTime.now().minusDays(daysBefore).toInstant());
-
-            messages.add("days before: %s, date to search: %s".formatted(daysBefore, date.toString()));
-            repository.deleteBySeenIsTrueAndReceivedDateBefore(date);
-            messages.add("after cleaning, count %s".formatted(repository.count()));
+            if (getFromParam(parameters) instanceof Tuple2<Long, Date>(Long daysBefore, Date date)) {
+                messages.add("days before: %s, date to search: %s".formatted(daysBefore, date.toString()));
+                repository.deleteBySeenIsTrueAndReceivedDateBefore(date);
+                messages.add("after cleaning, count %s".formatted(repository.count()));
+            }
             return resultBuilder.finishedDate(new Date()).messages(messages).build();
         } catch (Exception e) {
             log.error("error while executing action", e);
